+++ packages/client/src/app/providers/layout/MainLayout/MainLayout.module.scss
