+++ packages/client/src/app/providers/layout/MainLayout/MainLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router'
 import styles from './MainLayout.module.scss'
-import { Header } from '../../components/Header'
+import { Header } from '@/widgets/Header'
 export const MainLayout = () => {
   return (
     <>
