+++ packages/client/src/app/providers/layout/ProtectedLayout/ProtectedLayout.tsx
