+++ packages/client/src/app/providers/layout/PublicLayout/PublicLayout.tsx
