+++ packages/client/src/app/providers/layout/index.ts
@@ -1,0 +1,5 @@
+import { MainLayout } from './MainLayout/MainLayout'
+import { ProtectedLayout } from './ProtectedLayout/ProtectedLayout'
+import { PublicLayout } from './PublicLayout/PublicLayout'
+
+export { MainLayout, ProtectedLayout, PublicLayout }
