+++ packages/client/src/app/providers/styles/commonTokens.scss
