+++ packages/client/src/entities/User/ui/RegisterForm/ui/RegisterForm.module.scss
