+++ packages/client/src/main.tsx
@@ -1,5 +1,5 @@
 import ReactDOM from 'react-dom/client'
-import './styles/normilize.scss'
+import '@/app/providers/styles/normilize.scss'
 import App from './app/App'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
