+++ packages/client/src/shared/constants/auth.ts
@@ -1,15 +1,3 @@
-export interface IRegisterFormValues {
-  username: string
-  email: string
-  first_name: string
-  phone: string
-  second_name: string
-  password: string
-  confirm_password: string
-}
-
-export type RegisterFormField = keyof IRegisterFormValues
-
 export enum AuthFormPlaceholders {
   USERNAME = 'Имя пользователя',
   EMAIL = 'Email',
