+++ packages/client/src/shared/context/ThemeContext/index.ts
@@ -1,0 +1,2 @@
+export { ThemeContext } from './ThemeContext'
+export { useTheme } from './useTheme'
