+++ packages/client/src/shared/context/ThemeContext/useTheme.tsx
@@ -1,0 +1,4 @@
+import { useContext } from 'react'
+import { ThemeContext } from './ThemeContext'
+
+export const useTheme = () => useContext(ThemeContext)
