+++ packages/client/src/shared/types/auth.ts
@@ -1,0 +1,11 @@
+export interface IRegisterFormValues {
+  username: string
+  email: string
+  first_name: string
+  phone: string
+  second_name: string
+  password: string
+  confirm_password: string
+}
+
+export type RegisterFormField = keyof IRegisterFormValues
